Apply "modern c" fixes.
Some errno.h / strerror() fixes that I had applied in the distant
past weren't in the tarball I had on a web server.  At VCFmw 2021,
an exhibitor was trying to build the program for use with machines
in his exhibit, and we worked out the cause and re-applied the fixes.

I don't recall the background of the other three files' changes.